Move package from pkg to internal
